Adapt legacy providers

+++ src/extension-scripts/background.js
@@ -1,8 +1,15 @@
 import OCSproviders from "/src/data/providers.json";
 import OCSfunctions from "/src/data/functions.json";
 import OCSoptions from "/src/data/options.json";
-import { sortByPosition, adaptLegacyProvider } from "/src/modules/Utilities";
+import { sortByPosition } from "../modules/Utilities";
 import { get, set } from "../modules/Utilities";
+import legacyData from "../data/legacyData.js";
+import { adaptLegacyObject, isLegacyData } from "../modules/AdaptLegacyData";
+
+const devConfig = {
+  setLegacy: false,
+  clear: true,
+};
 
 const sw_log = (...args) => {
   console.log("service worker:", ...args);
@@ -14,20 +21,35 @@ const defaults = {
   options: OCSoptions,
 };
 
-//todo: adapt for legacy
-// get(
-//   { providers: defaultProviders, options: defaultOptions },
-//   (result) =>
-//     adaptForLegacy(result, set);
-//     })
-// );
+if (devConfig.clear) {
+  chrome.storage.sync.clear(() => sw_log("Storage was cleared"));
+}
 
-// clear during dev
-chrome.storage.sync.clear(() => sw_log("Storage was cleared"));
+if (devConfig.setLegacy) {
+  set(legacyData, () => sw_log("Storage was set to legacyData"));
+} else {
+  // Check data in storage
+  get(defaults, (result) => {
+    // If data existed in storage, stored data is returned.
+    // If no data existed, defaults are returned
 
-// Check if data exists; if not: set to defaults
-get(defaults, (result) =>
-  set(result, () => {
-    sw_log("This object was set in storage:\n", result);
-  })
-);
+    let dataToStore = result;
+
+    // Check is the data is legacy data and needs to be converted
+    const isLegacy = isLegacyData(result);
+
+    if (isLegacy) {
+      const newProvidersArray =
+        adaptLegacyObject(result.providers) ?? OCSproviders; // Will use defaults if adaptation fails
+
+      dataToStore = {
+        providers: sortByPosition([...newProvidersArray, ...OCSfunctions]),
+        options: OCSoptions,
+      };
+    }
+
+    set(dataToStore, () => {
+      sw_log("This object was set in storage:\n", dataToStore);
+    });
+  });
+}
